requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
requests
motor
pymongo
python-jose[cryptography]
passlib[bcrypt]
groq
google-generativeai
huggingface_hub
python-multipart
authlib
itsdangerous
slowapi
cachetools
stripe
redis
sendgrid
pywebpush
feedparser==6.0.10
legacy-cgi==2.6.3
pwnedpasswords
requests==2.31.0